+++ pkg/podgrouper/pod_controller.go
@@ -26,8 +26,7 @@ import (
 )
 
 const (
-	jobIdAnnotationForPod = "runai-job-id"
-	controllerName        = "pod-grouper"
+	controllerName = "pod-grouper"
 
 	rateLimiterBaseDelay = time.Second
 	rateLimiterMaxDelay  = time.Minute
@@ -152,28 +151,15 @@ func (r *PodReconciler) addPodGroupAnnotationToPod(ctx context.Context, pod *v1.
 		pod.Annotations = map[string]string{}
 	}
 
-	reconcile := false
 	value, found := pod.Annotations[constants.PodGroupAnnotationForPod]
-	if !found || value != podGroup {
-		logger.V(1).Info("Reconciling podgroup annotation for pod", "pod",
-			fmt.Sprintf("%s/%s", pod.Namespace, pod.Name), "old", value, "new", podGroup)
-		reconcile = true
-	}
-
-	value, found = pod.Annotations[jobIdAnnotationForPod]
-	if !found || value != jobID {
-		logger.V(1).Info("Reconciling jobId annotation for pod", "pod",
-			fmt.Sprintf("%s/%s", pod.Namespace, pod.Name), "old", value, "new", jobID)
-		reconcile = true
-	}
-
-	if !reconcile {
+	if found && value == podGroup {
 		return nil
 	}
+	logger.V(1).Info("Reconciling podgroup annotation for pod", "pod",
+		fmt.Sprintf("%s/%s", pod.Namespace, pod.Name), "old", value, "new", podGroup)
 
 	newPod := pod.DeepCopy()
 	newPod.Annotations[constants.PodGroupAnnotationForPod] = podGroup
-	newPod.Annotations[jobIdAnnotationForPod] = jobID
 
 	return r.Client.Patch(ctx, newPod, client.MergeFrom(pod))
 }