+++ pkg/podgrouper/podgrouper/plugins/constants/constants.go
@@ -8,7 +8,6 @@ const (
 	ProjectLabelKey    = "project"
 	PriorityLabelKey   = "priorityClassName"
 	UserLabelKey       = "user"
-	JobIdKey           = "runai/job-id"
 
 	BuildPriorityClass     = "build"
 	TrainPriorityClass     = "train"