+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fixed pod status scheduled race condition between the scheduler and the pod binding
 - Removed redundant `replicas` key for binder from `values.yaml` as it is not used and not supported
 
+### Removed
+- Removed `runai-job-id` and `runai/job-id` annotations from pods and podgroups
 
 ## [v0.5.1] - 2025-05-20